refactor book status logic
